+++ test/text_comparison_score_codespark_test.dart
@@ -7,7 +7,8 @@ void main() {
       String string1 = "kitten";
       String string2 = "sitting";
 
-      double result = TextComparisonScore.matchPercentage(string1, string2);
+      double result = TextComparisonScore.calculateScore(string1, string2,
+          algorithm: ComparisonAlgorithm.jaroWinkler);
 
       expect(result, 57.14285714285714);
     });
@@ -17,7 +18,7 @@ void main() {
       String string1 = "flutter";
       String string2 = "flutter";
 
-      double result = TextComparisonScore.matchPercentage(string1, string2);
+      double result = TextComparisonScore.calculateScore(string1, string2);
 
       expect(result, 100.0);
     });
@@ -28,7 +29,7 @@ void main() {
       String string1 = "apple";
       String string2 = "orange";
 
-      double result = TextComparisonScore.matchPercentage(string1, string2);
+      double result = TextComparisonScore.calculateScore(string1, string2);
 
       expect(result, 0.0);
     });
@@ -37,7 +38,7 @@ void main() {
       String string1 = "kitten";
       String string2 = "";
 
-      double result = TextComparisonScore.matchPercentage(string1, string2);
+      double result = TextComparisonScore.calculateScore(string1, string2);
 
       expect(result, 0.0);
     });
@@ -46,7 +47,7 @@ void main() {
       String string1 = "";
       String string2 = "";
 
-      double result = TextComparisonScore.matchPercentage(string1, string2);
+      double result = TextComparisonScore.calculateScore(string1, string2);
 
       expect(result, 100.0);
     });
@@ -55,7 +56,7 @@ void main() {
       String string1 = "dart";
       String string2 = "start";
 
-      double result = TextComparisonScore.matchPercentage(string1, string2);
+      double result = TextComparisonScore.calculateScore(string1, string2);
 
       expect(result, 75.0);
     });
@@ -64,7 +65,7 @@ void main() {
       String string1 = "flame";
       String string2 = "frame";
 
-      double result = TextComparisonScore.matchPercentage(string1, string2);
+      double result = TextComparisonScore.calculateScore(string1, string2);
 
       expect(result, 80.0);
     });