+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2024-08-25
+
+### Added
+- Added the `Jaro-Winkler Distance` algorithm to the `TextComparisonScore` class.
+  - This algorithm provides a string similarity score with more favorable ratings for strings that share a common prefix.
+  - Added support for comparing strings using Jaro-Winkler Distance and calculating match percentage.
+- Updated `README.md` with usage examples and details on when to use the Jaro-Winkler algorithm.
+
 ## [0.0.4] - 2024-08-13
 
 ### Added